akka-apps: cleanup on the camera/listen/screen permission calls
